Hax to run uWSGI tests.
Extends the uWSGI test suite to spawn uWSGI, then run that same test
file as a WSGI app, then make an HTTP request against that running
instance, asserting the response code, where the response code is 204 if
the uWSGI tests pass and 500 if they fail.

This is done because there does not seem to be a way to mock the uWSGI
Python module beyond running uWSGI itself.

The uWSGI tests must be run within a virtual env, which is used to
determine the location of pytest and uwsgi.