Switch multi-process support to custom data store.
This uses an mmaped file to store the data.
Fsync is not called, and concurrent readers are permitted (which
some shelve backends did not permit).

Microbenchmarks indicate 1.2-1.5us per instrumentation event.
This compares to 3-13ms from the shelve solution, so 3 orders
of magnitude better.

The 1.2-1.5us is about double normal multi-process instrumentation, due
to the 2nd mutex required. Python doesn't offer an RWLock which would
let this be reduced.

Overall this performance is acceptable, as if someone was doing
something extremely performance critical with instrumentation they
wouldn't be using Python.