Add describe  and autodescribe to detect dupes.
This works by adding an optional describe method to
collectors, which returns data in the same format as collect (though
hopefully without the samples).
If present it is called at registration time.

If describe is not present and auto_describe=True is set on the
registry, then collect is called instead. This is enabled by default on
the default registry, but disabled elsewhere. There's not much
point doing extra checks for a transient registry used to
push to a pushgateway, particularly if collection is expensive.