Fixes Histogram decorator usage (#140)
Without creating a new instance of `_HistogramTimer` for each call, the internal timer is reused across executions of the instrumented function.  This results in erroneous timings.  The solution is to create a new `_HistogramTimer` for each execution of the decorated function.  This provides the same behavior as using the `Histogram.time()` as a decorator within the instrumented function.