Refactor internal sample to use tuple for labels
This preserves the order of labels at all times. The only place a dict
of labels was used in the public api is in get_sample_value(), where we
convert to dict for the comparision, which seems reasonable for
a testing helper.

Originally tried just using OrderedDict, for the best of both worlds,
but sadly python 2.6 doesn't have OrderedDict.

Signed-off-by: Simon Davy <simon.davy@canonical.com>