using worker_id instead of pid will produce fewer db files
Signed-off-by: fangjianfeng <fangjianfeng@rightpaddle.com>

using worker_id instead of pid will produce fewer db files

As the program is restarted frequently, more and more db files will be generated. These files are cumbersome to manage, and aggregate metrics also consume resources.

If the worker id is used as the db file naming method, as long as the number of wokers is determined and the metric type is unchanged, the number of generated db files will not increase, but the values ​​in the file are changing.