Fix references to non-existant metric fields
I found this bug by writing a custom collector that returns a list of `Gauge` and `Counter` (both of these classes extend `MetricWrapperBase`). `MetricWrapperBase`, `Gauge`, and `Counter` do not have the property `name` but rather `_name`. Same goes for `_type`, `_documentation`. The error about missing attribute is thrown when implementing a custom collector with no `describe` function, in which case the code defaults to using the `collect` function. An easy workaround is writing a `describe` function that returns an empty list, but in any case the code here should be corrected to reference the correct variable names.

Signed-off-by: Dan Bunker <dbunked@gmail.com>