Document implications of negative observations
Fixes #638.

It's of course weird to talk abou negative values of an
"amount". However, for the Gauge, which quite regularly will receive
negative increments, the word "amount" is used, too, so I didn't want
to change it everywhere to "value" or something.

AFAIK the only place where negative values are rejected is indeed the
Counter.

Signed-off-by: beorn7 <beorn@grafana.com>