Allow adding arbitrary labels to metrics when generating output
It's useful to be able to apply a label across all metrics, for
example when using a multi-processing framework to apply a
label for worker ID to all metrics (including from builtin
collector types). This change alows passing a dictionary of
labels and values to generate_latest which will then be applied
to all generated metrics.

Signed-off-by: Will Newton <will.newton@gmail.com>