Allow Prometheus parser to handle missing space
If there is no delimiter separating the labels and value when parsing
the Prometheus format the parser currently will chop off the first digit
of the value. Instead, allow a missing delimiter in the Prometheus
format to align with what the Go parser does.

In addition, add a test to specifically check that a delimiter is still
required in OpenMetrics.

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>