Use License Expressions in pyproject.toml
With the release of PEP-639 the best practice for specifying the license
is now to use a license expression in the license field and specify any
license files in license-files rather than the table-based approach from
PEP-621. Including the license in the classifiers is also no longer
allowed when using PEP-639 and has been removed.

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>