Add support to `write_to_textfile` for custom tmpdir (#1115)
* Add support to write_to_textfile for custom tmpdir

While the try/except block does prevent most of the temp files from
persisting, if there is a non catchable exception, those temp files
continue to pollute the directory. Optionally set the temp
directory would let us write to something like /tmp, so the target
directory isn't polluted

Signed-off-by: Aaditya Dhruv <aadityadhruv@mailbox.org>

* Modify write_to_textfile to ensure tmpdir is on same filesystem

The tmpdir must be on the same filesystem to ensure an atomic operation
takes place. If this is not enforced, there could be partial writes
which can lead to partial/incorrect metrics being exported

Signed-off-by: Aaditya Dhruv <aadityadhruv@mailbox.org>

---------

Signed-off-by: Aaditya Dhruv <aadityadhruv@mailbox.org>