fix: Use `asyncio.new_event_loop()` to create event loop for tests (#1138)
* fix: Use `asyncio.new_event_loop()` to create event loop for tests

Replace the use of `asyncio.get_event_loop()` with more appropriate
`asyncio.new_event_loop()` to create event loops for testing.
The former used to be a wrapper that either returned the currently
running event loop or created a new one, but the latter behavior was
deprecated and removed in Python 3.14.  Since the tests are always run
in a synchronous context, and they always run the obtained event loop
to completion, just always create a new event loop.

Fixes #1137

Signed-off-by: Michał Górny <mgorny@gentoo.org>

* fix: Remove obsolete asgiref pin

Remove the `asgiref` pin linked to #1020.  I can't reproduce the issue
anymore with the current `asgiref` versions, and the pin actually
breaks the tests with the `asyncio` event loop fixes.

Signed-off-by: Michał Górny <mgorny@gentoo.org>

---------

Signed-off-by: Michał Górny <mgorny@gentoo.org>