Add an AIOHTTP exporter (#1139)
* Always run the asgi tests

Since the client now requires a minimum of Python 3.9, we don't need to
have this feature gate in place any more

Signed-off-by: Lexi Robinson <lexi@lexi.org.uk>

* Add an AIOHTTP exporter

Unfortunately the AIOHTTP library doesn't support ASGI and apparently
has no plans to do so which makes the ASGI exporter not suitable for
anyone using it to run their python server.

Where possible this commit follows the existing ASGI implementation and
runs the same tests for consistency.

Signed-off-by: Lexi Robinson <lexi@lexi.org.uk>

---------

Signed-off-by: Lexi Robinson <lexi@lexi.org.uk>